FAC-963:impl:Move ApproveAllAssessmentInsights use case to insight package
