+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/in/rest/insight/ApproveAllAssessmentInsightsRestController.java
@@ -1,9 +1,9 @@
-package org.flickit.assessment.core.adapter.in.rest.assessment;
+package org.flickit.assessment.core.adapter.in.rest.insight;
 
 import lombok.RequiredArgsConstructor;
 import org.flickit.assessment.common.config.jwt.UserContext;
-import org.flickit.assessment.core.application.port.in.assessment.ApproveAllAssessmentInsightsUseCase;
-import org.flickit.assessment.core.application.port.in.assessment.ApproveAllAssessmentInsightsUseCase.Param;
+import org.flickit.assessment.core.application.port.in.insight.ApproveAllAssessmentInsightsUseCase;
+import org.flickit.assessment.core.application.port.in.insight.ApproveAllAssessmentInsightsUseCase.Param;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
