+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/service/insight/ApproveAllAssessmentInsightsService.java
@@ -1,9 +1,9 @@
-package org.flickit.assessment.core.application.service.assessment;
+package org.flickit.assessment.core.application.service.insight;
 
 import lombok.RequiredArgsConstructor;
 import org.flickit.assessment.common.application.domain.assessment.AssessmentAccessChecker;
 import org.flickit.assessment.common.exception.AccessDeniedException;
-import org.flickit.assessment.core.application.port.in.assessment.ApproveAllAssessmentInsightsUseCase;
+import org.flickit.assessment.core.application.port.in.insight.ApproveAllAssessmentInsightsUseCase;
 import org.flickit.assessment.core.application.port.out.assessmentinsight.ApproveAssessmentInsightPort;
 import org.flickit.assessment.core.application.port.out.attributeinsight.ApproveAttributeInsightPort;
 import org.flickit.assessment.core.application.port.out.subjectinsight.ApproveSubjectInsightPort;
